+++ src/smart-components/access-management/UserGroupsTable.tsx
@@ -1,12 +1,20 @@
-import React, { useEffect, useCallback, useMemo, Fragment } from 'react';
+import React, { useEffect, useCallback, useMemo, useState, Fragment } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useDataViewSelection, useDataViewPagination } from '@patternfly/react-data-view/dist/dynamic/Hooks';
 import { BulkSelect, BulkSelectValue } from '@patternfly/react-component-groups/dist/dynamic/BulkSelect';
 import { DataView } from '@patternfly/react-data-view/dist/dynamic/DataView';
 import { DataViewToolbar } from '@patternfly/react-data-view/dist/dynamic/DataViewToolbar';
 import { DataViewTable } from '@patternfly/react-data-view/dist/dynamic/DataViewTable';
-import { ButtonVariant, Pagination, Tooltip } from '@patternfly/react-core';
-import { ActionsColumn } from '@patternfly/react-table';
+import { 
+  ButtonVariant, 
+  EmptyState, 
+  EmptyStateBody, 
+  EmptyStateHeader, 
+  EmptyStateIcon, 
+  Pagination, 
+  Tooltip 
+} from '@patternfly/react-core';
+import { ActionsColumn, TableVariant } from '@patternfly/react-table';
 import { mappedProps } from '../../helpers/shared/helpers';
 import { RBACStore } from '../../redux/store';
 import { useSearchParams } from 'react-router-dom';
@@ -15,8 +23,9 @@ import { formatDistanceToNow } from 'date-fns';
 import { FormattedMessage, useIntl } from 'react-intl';
 import messages from '../../Messages';
 import { Group } from '../../redux/reducers/group-reducer';
-import { DataViewTrObject, EventTypes, useDataViewEventsContext } from '@patternfly/react-data-view';
-import { WarningModal } from '@patternfly/react-component-groups';
+import { DataViewTrObject, DataViewState, EventTypes, useDataViewEventsContext } from '@patternfly/react-data-view';
+import { SearchIcon } from '@patternfly/react-icons';
+import { SkeletonTable, WarningModal } from '@patternfly/react-component-groups';
 
 const COLUMNS: string[] = ['User group name', 'Description', 'Users', 'Service accounts', 'Roles', 'Workspaces', 'Last modified'];
 
@@ -48,6 +57,7 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
   const [isDeleteModalOpen, setIsDeleteModalOpen] = React.useState(false);
   const [currentGroups, setCurrentGroups] = React.useState<Group[]>([]);
   const dispatch = useDispatch();
+  const [activeState, setActiveState] = useState<DataViewState | undefined>(DataViewState.loading);
   const intl = useIntl();
   const { trigger } = useDataViewEventsContext();
 
@@ -56,9 +66,10 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
     setIsDeleteModalOpen(!isDeleteModalOpen);
   };
 
-  const { groups, totalCount } = useSelector((state: RBACStore) => ({
+  const { groups, totalCount, isLoading } = useSelector((state: RBACStore) => ({
     groups: state.groupReducer?.groups?.data || [],
     totalCount: state.groupReducer?.groups?.meta.count || 0,
+    isLoading: state.groupReducer?.isLoading,
   }));
 
   let pagination;
@@ -101,6 +112,14 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
       count: totalCount || 0,
     });
   }, [fetchData, page, perPage]);
+
+  useEffect(() => {
+    if (isLoading) {
+      setActiveState(DataViewState.loading);
+    } else {
+      totalCount === 0 ? setActiveState(DataViewState.empty) : setActiveState(undefined);
+    }
+  }, [totalCount, isLoading]);
 
   useEffect(() => {
     onChange?.(selected);
@@ -194,6 +213,24 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
     />
   );
 
+  const empty = (
+    <EmptyState>
+      <EmptyStateHeader
+        titleText={intl.formatMessage(messages.userGroupsEmptyStateTitle)}
+        headingLevel="h4"
+        icon={<EmptyStateIcon icon={SearchIcon} />}
+      />
+      <EmptyStateBody>
+        <FormattedMessage
+          {...messages['userGroupsEmptyStateSubtitle']}
+          values={{
+            br: <br />,
+          }}
+        />
+      </EmptyStateBody>
+    </EmptyState>
+  );
+
   return (
     <Fragment>
       {isDeleteModalOpen && (
@@ -221,7 +258,7 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
           />
         </WarningModal>
       )}
-      <DataView ouiaId={ouiaId} selection={selection}>
+      <DataView ouiaId={ouiaId} selection={selection} activeState={activeState}>
         <DataViewToolbar
           ouiaId={`${ouiaId}-header-toolbar`}
           bulkSelect={
@@ -258,7 +295,18 @@ const UserGroupsTable: React.FunctionComponent<UserGroupsTableProps> = ({
           }
           pagination={React.cloneElement(paginationComponent, { isCompact: true })}
         />
-        <DataViewTable variant="compact" aria-label="Users Table" ouiaId={`${ouiaId}-table`} columns={COLUMNS} rows={rows} />
+        {isLoading ? (
+          <SkeletonTable rowsCount={10} columns={COLUMNS} variant={TableVariant.compact} />
+        ) : (
+          <DataViewTable 
+            variant="compact" 
+            aria-label="Users Table" 
+            ouiaId={`${ouiaId}-table`} 
+            columns={COLUMNS} 
+            rows={rows} 
+            states={{ empty }} 
+          />
+        )}
         <DataViewToolbar ouiaId={`${ouiaId}-footer-toolbar`} pagination={paginationComponent} />
       </DataView>
     </Fragment>